ìvalidation

+++ src/main/java/shop/mtcoding/blog/board/BoardController.java
@@ -1,6 +1,5 @@
 package shop.mtcoding.blog.board;
 
-import jakarta.persistence.Query;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpSession;
 import lombok.RequiredArgsConstructor;
@@ -8,6 +7,8 @@ import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
+import shop.mtcoding._core.errors.Exception403;
+import shop.mtcoding._core.errors.Exception404;
 import shop.mtcoding.blog.user.User;
 
 import java.util.List;
@@ -20,23 +21,37 @@ public class BoardController {
     private final HttpSession session;
 
     @PostMapping("/board/{id}/delete")
-    public String deleteById (@PathVariable(name = "id") Integer id){
+    public String deleteById(@PathVariable(name = "id") Integer id) {
+        User sessionUser = (User) session.getAttribute("sessionUser");
+        Board board = boardRepository.findById(id);
+
+        if(sessionUser.getId() != board.getUser().getId()){
+            throw new Exception403("게시글을 삭제할 권한이 없습니다");
+        }
+
         boardRepository.deleteById(id);
         return "redirect:/";
     }
 
 
     @PostMapping("board/save")
-    public String save (BoardRequest.SaveDTO reqDTO){
+    public String save(BoardRequest.SaveDTO reqDTO) {
         User sessionUser = (User) session.getAttribute("sessionUser");
         boardRepository.save(reqDTO.toEntity(sessionUser));
         return "redirect:/";
     }
 
     @PostMapping("/board/{id}/update")
-    public String updateById (@PathVariable(name = "id") Integer id, BoardRequest.UpdateDTO reqDTO){
+    public String updateById(@PathVariable(name = "id") Integer id, BoardRequest.UpdateDTO reqDTO) {
+        User sessionUser = (User) session.getAttribute("sessionUser");
+        Board board = boardRepository.findById(id);
+
+        if(sessionUser.getId() != board.getUser().getId()){
+            throw new Exception403("게시글을 수정할 권한이 없습니다");
+        }
+
         boardRepository.updateById(id, reqDTO.getTitle(), reqDTO.getContent());
-        return "redirect:/board/"+id;
+        return "redirect:/board/" + id;
     }
 
     @GetMapping("/")
@@ -59,13 +74,26 @@ public class BoardController {
 
     @GetMapping("/board/{id}")
     public String detail(@PathVariable(name = "id") Integer id, HttpServletRequest request) {
+        User sessionUser = (User) session.getAttribute("sessionUser");
         Board board = boardRepository.findByIdJoinUser(id);
+        boolean isOnwer = false;
+        if (sessionUser != null) {
+            if (sessionUser.getId() == board.getUser().getId()) {
+                isOnwer = true;
+            }
+        }
         request.setAttribute("board", board);
         return "board/detail";
     }
+
     @GetMapping("/board/{id}/update-form")
-    public String updateForm(@PathVariable(name ="id") Integer id, HttpServletRequest request) {
+    public String updateForm(@PathVariable(name = "id") Integer id, HttpServletRequest request) {
         Board board = boardRepository.findById(id);
+
+        if(board == null){
+            throw new Exception404("해당 게시글을 찾을 수 없습니다");
+        }
+
         request.setAttribute("board", board);
         return "board/update-form";
     }
